Use bytes instead of ints for loop variables where possible

+++ Chords.cpp
@@ -15,11 +15,11 @@ chord enchord(size_t len, int * notes) {
 } */
 
 // Compose a chord out of notes, the exciting and dangerous variadic version!
-Chord enchord(int numnotes, ...) {
+Chord enchord(byte numnotes, ...) {
     Chord outchord = 0;
     va_list args;
     va_start(args, numnotes);
-    for (int i = 1; i <= numnotes; i++) {
+    for (byte i = 1; i <= numnotes; i++) {
         int val = va_arg(args, int);
         outchord |= Notes[val];
     }
@@ -29,7 +29,7 @@ Chord enchord(int numnotes, ...) {
 
 // Apply a function to each note of a chord
 void chordmap(Chord input, void (*func)(int note)) {
-    int i = 1;
+    byte i = 1;
     long comparison = input;
     while (comparison > 0) {
         if (Notes[i] & comparison) {
